Ensure the parent process is always killed
If you try to run the test suite but do not have a redis instance
running, the parent process that was supposed to be killed will never be
as the line with posix_kill() is never reached.
This results in the test suite hanging forever at the end.
In this patch, throwables thrown during store instantiation are caught
(they were not caught before), and the parent gets killed as it should
in the catch statement.